Only append bar to the first body element
In the case where one is using an svg diagram with a foreign object element which contains it's own self-contained html DOM, the progress bar was being applied to all of them. Resulting in many loading bars.
This patch restricts the selector to only find one element, which will the first.